+++ react_staging/02_src_hello_react/components/Hello/index.jsx
@@ -1,0 +1,11 @@
+import React,{Component} from "react";
+import hello from './index.module.css'
+export default class Hello extends Component{
+    render(){
+        return(
+            <div className={hello.title}>
+                Hello,React!!!
+            </div>
+        )
+    }
+}
