+++ react_staging/02_src_hello_react/components/Welcome/index.css
@@ -1,0 +1,3 @@
+.title {
+    background-color: yellow;
+}
