+++ react_staging/02_src_hello_react/components/Welcome/index.jsx
@@ -1,0 +1,11 @@
+import React,{Component} from "react";
+import './index.css'
+export default class Welcome extends Component{
+    render(){
+        return(
+            <h2 className="title">
+                Welcome!
+            </h2>
+        )
+    }
+}
