+++ react_staging/02_src_hello_react/components/css/index.css
@@ -1,0 +1,4 @@
+.title {
+  background-color: red;
+  color: aquamarine;
+}
