+++ react_staging/02_src_hello_react/index.js
@@ -1,0 +1,9 @@
+// 引入react核心库
+import React from 'react'
+// 引入ReactDOM
+import ReactDOM from 'react-dom'
+// 引入App组件
+import App from './App'
+
+// 渲染App到页面
+ReactDOM.render(<App/>,document.getElementById('root'))
